feat(slo): bugfix for int?

+++ internal/resources/slo/resource_slo.go
@@ -568,7 +568,7 @@ func packAlerting(tfAlerting map[string]interface{}) slo.SloV00Alerting {
 		lf := failures.([]interface{})
 		if len(lf) > 0 {
 			lf2 := lf[0].(map[string]interface{})
-			tfAdvancedOptions.SetMinFailures(int64(lf2["min_failures"].(float64)))
+			tfAdvancedOptions.SetMinFailures(int64(lf2["min_failures"].(int)))
 		}
 	}
